model/requirements.txt
requests
bs4
selenium
pillow
tenacity
argparse
numpy
torch
torchvision
accelerate
transformers
diffusers
PyWavelets
scikit-learn
matplotlib